requirements.txt
Flask
pandas
openpyxl
reportlab
Werkzeug
python-dotenv
xlrd
arabic-reshaper
python-bidi
pytest
